Change function to use list comprehention
Changed function to use list comprehention instead of appending to the
list one element at a time.

Change how some functions work

In `BlockCache.get_missing_in_range` used list comprehention instead of
manually appending to a new list in a loop.

In `OrderBook._get_order_ids` used `map` instead of manually yielding
in a loop.

In `MatchPriorityQueue` used a `namedtuple` as the element of the internal
`queue` list. Used `any` with a generator expression in `contains_order`
instead of the loop with `_, _, other_order_id, _` destructuring,
utilizing the introduced `namedtuple`. In `insert` utilized the new
`namedtuple` in the internal `cmp_items` function for maintainability
and readability. Also in `insert` instead of creating a new sorted list
for `self.queue` the queue is sorted in-place.