Use `next` with default value `None`
Used `next` with a default value of `None` instead of using `try-except`
on `StopIteration`.

Removed `u` prefix from string. The prefix is superfluous in python 3.

Instead of checking `current_version` against individual numbers it is
checked against a tuple of values.

Change parameterized `super()` to empty `super()`

Instead of calling `super()` with explicit parameter, an empty call to
it is made instead.