Change ValueError to TypeError when appropriate
Changed the raising of ValueError to TypeError when appropriate. Namely
when an argument's type is checked. Likewise changed the tests to match
the new exceptions thrown.

In `bitcoinlib_main.py` moved some software structures out of a file's
`with` statement in order to not hold the file open longer than
necessary. In addition restructured a for-loop in a minor way in order
to decrease the indentation level of the loop's body.